% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educ_data}
\alias{educ_data}
\title{Data as example}
\format{
A dataframe with two variables and 4 rows:
\describe{
\item{Education}{Education score}
\item{Age}{Age of person}
}
}
\source{
{Created as example}
}
\usage{
educ_data
}
\description{
Contaisn two variables
}
\keyword{datasets}
